chore(codegen): refactor
