+++ packages/codegen-tests/package.json
@@ -48,7 +48,7 @@
 		"node": ">=18"
 	},
 	"scripts": {
-
+		"buildAndTest": "cd ../codegen && pnpm build && cd ../codegen-tests && pnpm i && pnpm test",
 		"test": "vitest run --coverage  --outputFile=../../test_reports/codegen.md",
 		"lint": "eslint .",
 		"format": "eslint --fix ."
