--- packages/codegen/src/core/modules/JavaScript/getDefaultPropertyContext.ts
@@ -1,14 +1,0 @@
-import { getReferenceString } from './shorties';
-
-export function getDefaultPropertyContext(path: string, indetifier: string, expression?: string) {
-	const fullPath = getReferenceString(path, indetifier);
-
-	return `(function(){
-						if(${path} !== null && ${fullPath} !== undefined && ${fullPath} !== null) {
-							return ${path}['${indetifier}']
-						}
-							else {
-								return ${expression ?? 'null'}
-							}
-					}())`;
-}
